Renamed packages. Added comments.
