+++ src/main/java/com/umermansoor/trafficdistributor/config/Configuration.java
@@ -1,4 +1,4 @@
-package com.umermansoor.trafficdistributor;
+package com.umermansoor.trafficdistributor.config;
 
 
 public class Configuration {
