+++ src/main/java/com/umermansoor/trafficdistributor/net/IncomingConnectionManager.java
@@ -1,31 +1,35 @@
 package com.umermansoor.trafficdistributor.net;
 
-import com.umermansoor.trafficdistributor.Configuration;
+import com.umermansoor.trafficdistributor.config.Configuration;
 import org.slf4j.Logger;
 
 import java.net.ServerSocket;
 import java.net.Socket;
 import java.util.Queue;
+import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 
 /**
- * Handles incoming connections.
+ * Manages incoming TCP connections from clients.
+ *
+ * @author umer mansoor
  */
 public class IncomingConnectionManager implements Runnable {
     private static final Logger logger = org.slf4j.LoggerFactory.getLogger(IncomingConnectionManager.class);
     private final Queue<String> centralQueue;
     private final ExecutorService pool = Executors.newFixedThreadPool(Configuration.MAX_CLIENT_CONNECTIONS);
+    private final CountDownLatch serverStartedSignal;
     ServerSocket server;
 
-    public IncomingConnectionManager(Queue<String> q) {
+    public IncomingConnectionManager(Queue<String> q, CountDownLatch l) {
         centralQueue = q;
+        serverStartedSignal = l;
 
     }
 
     public void run() {
         startServer();
-
     }
 
     private void startServer() {
@@ -36,6 +40,8 @@ public class IncomingConnectionManager implements Runnable {
             logger.error("failed to bind to port {}. {}", Configuration.LISTENING_PORT, e.toString());
             return;
         }
+
+        serverStartedSignal.countDown();
 
         logger.debug("listening on port {} for clients.", Configuration.LISTENING_PORT);
 
@@ -65,6 +71,13 @@ public class IncomingConnectionManager implements Runnable {
         }
     }
 
+    /**
+     * Normally to cancel a task(Runnable), we send am interrupt signal. But
+     * the call to `ServerSocket.accept()` ignores interrupt and blocks until a
+     * connection request is received. As a work around, this method closes the
+     * underlying `Socket` forcing the `accept()` method to throw an exception,
+     * giving the current thread a chance to process the interrupt request.
+     */
     public synchronized void cancelTask() {
         if (server != null) {
             try {
